% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fin1933-data.R
\docType{data}
\name{fin1933}
\alias{fin1933}
\title{Input data for Finland 1933}
\format{
An object of class \code{data.frame} with 23 rows and 7 columns.
}
\source{
Under-Five Mortality Database (U5MD)
}
\usage{
fin1933
}
\description{
Input data for Finland 1933
}
\references{
\url{https://repository.upenn.edu/psc_publications/54/}.
}
\keyword{datasets}
