% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_format.R
\name{format_data}
\alias{format_data}
\title{format_data}
\usage{
format_data(lower_age, upper_age, rate, type, sex, fit, weight)
}
\arguments{
\item{lower_age}{Lower boundary of the age interval in days for each input (see details below for values).}

\item{upper_age}{Upper boundary of the age interval in days for each input (see details below for values).}

\item{rate}{Mortality rate for each input.}

\item{type}{Character: 'qx' or 'mx'. Type of rate for each input.}

\item{sex}{Character: 'female', 'male', or 'total'. Same for all inputs.}

\item{fit}{Character: 'match' or 'min'. Needed only when there are more than 2 inputs. Select the input to be matched using 'match'. Use 'min' for all the other inputs.}

\item{weigth}{Optional when there are more than 2 inputs: weight for inputs whose root mean square error is minimized. Use 0 or NA for the matched input. The sum of weights must be equal to 1.}
}
\value{
\code{format_data} returns a data frame with the formated mortality inputs ready for the implementation of \code{lagrange5q0}.
}
\description{
\code{format_data} prepares and verifies mortality inputs for the implementation of \code{lagrange5q0}.
}
\details{
Possible boudaries for the age intervals (1 year = 365.25 days, 1 month = 30.4375 days):\tabular{ll}{
   Age \tab in Days \cr
   0 \tab 0 \cr
   7 days \tab 7 \cr
   14 days \tab 14 \cr
   21 days \tab 21 \cr
   28 days \tab 28 \cr
   2 months \tab 60.8750 \cr
   3 months \tab 91.3125 \cr
   4 months \tab 121.7500 \cr
   5 months \tab 152.1875 \cr
   6 months \tab 182.6250 \cr
   7 months \tab 213.0625 \cr
   8 months \tab 243.5000 \cr
   9 months \tab 273.9375 \cr
   10 months \tab 304.3750 \cr
   11 months \tab 334.8125 \cr
   12 months \tab 365.2500 \cr
   15 months \tab 456.5625 \cr
   18 months \tab 547.8750 \cr
   21 months \tab 639.1875 \cr
   2 years \tab 730.5000 \cr
   3 years \tab 1095.7500 \cr
   4 years \tab 1461.0000 \cr
   5 years \tab 1826.2500 \cr
}
}
\examples{
#One input: q(28d,5y) (Jordan 2015)
df <- format_data(
 rate      = 0.00804138,
 lower_age = 28,
 upper_age = 365.25*5,
 type      = "qx",
 fit       = "match",
 sex       = "total")

#Two inputs: q(0,1y) and q(1y,4y) (Belgium 1984)
df <- format_data(
 lower_age = c(0,365.25),
 upper_age = c(365.25,365.25*5),
 rate      = c(0.00996356, 0.00207788),
 type      = c("qx", "qx"),
 sex       = c("male", "male"))


#22 inputs + 1 match: q(x) (Finland 1933)
data(fin1933)
fin1933$weight <- c(fin1933$n[1:22]/(365.25*5), NA)
df <- format_data(
 lower_age = fin_1933$lower_age,
 upper_age = fin_1933$upper_age,
 rate      = fin_1933$rate,
 type      = fin_1933$type,
 sex       = fin_1933$sex,
 fit       = fin_1933$fit,
 weight    = fin_1933$weight)
}
