% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagrange5q0.R
\name{lagrange5q0}
\alias{lagrange5q0}
\title{lagrange5q0}
\usage{
lagrange5q0(data, k, model, region)
}
\arguments{
\item{data}{mortality inputs formated with \code{format_data}.}

\item{k}{constrained value of k (only when using a single mortality input).}

\item{model}{Character: 'A' (default) or 'B'.}

\item{region}{Character: 'Sub-Saharan Africa', 'Eastern Africa', 'Middle Africa', 'Western Africa', or 'South Asia' (only for Model B).}
}
\value{
\code{lagrange5q0} returns a list including six elements: the predicted values of q(5y) and k, a data frame with predicted values of q(x) and nMx, and the predicted value of 1a0, 4a1, and 5a0.

The boundaries of the age intervals are given in days (1 year = 365.25 days, 1 month = 30.4375 days).

Different age intervals are provived for the values of q(x), i.e. the cumulative probabilities of dying between 0 and age x, and for the values nMx,  i.e. the central death rates between x and x+n.

Two types of 95\% confidence intervals (CIs) are automatically estimated: (1)	CIs based on the patterns of prediction errors observed in the underlying data when using a single mortality input (k = 0), (2) CIs based on the uncertainty of k when minimizing the RMSE over a full series of 22 q(x).
}
\description{
\code{lagrange5q0} predicts a series of 22 cumulative probabilities of dying (q(x)) and mortality rates (nMx) for the first 5 years of life, based on one or more mortality inputs between 0 and 5 years.
}
\details{
\code{lagrange5q0} allows four types of mortality inputs:
\enumerate{
\item A single mortality input (either nqx or nMx) for any age interval. The value of k will be assumed equal to 0 (average outcome of the model) and the value of the mortality input will be matched exactly.
\item A single mortality input (either nqx or nMx) for any age interval and a value of k. Both the mortality input and the value of k and will be matched exactly.
\item Two mortality inputs (either nqx or nMx) for any age interval. For the purpose of estimating nax--the average number of years lived in the age interval x,x+n--with high precision, the proportion of infant deaths before 28 days or 3 months (z(28d) or z(3m)) can be used as second input along with another mortality input (either nqx or nMx) starting at age 0. The two mortality inputs will be matched exactly and the value of k will be estimated.
\item More than two mortality inputs (either nqx or nMx) for any age interval. One of them must be selected for matching. The selected input will be matched exactly, the root mean square error will be minimized over the remaining mortality inputs, and the value of k will be estimated.
}

The computation of the root mean square error (RMSE) can be weighted. When minimizing a series of q(x), i.e. cumulative probabilities of dying starting at age 0, we recommend using weights proportional to the length of the last age interval (see example 5 below).

Use \code{format_data} to prepare and verify the mortality inputs.

BETA VERSION: To use coefficients for sub-Saharan Africa and south Asia, specify 'B' for the parameter 'model'. In the case scenario of having a single mortality input and no information on the value of k, the user can specify the parameter 'region' ('Sub-Saharan Africa', 'Eastern Africa', 'Middle Africa', 'Western Africa', or 'South Asia') in order to benefit from a regional prior of k instead of assuming k = 0 (see example 6 below)
}
\examples{
#1. One input (k = 0): q(28d,5y) (Jordan 2015)
input <- format_data(
 rate      = 0.00804138,
 lower_age = 28,
 upper_age = 365.25*5,
 type      = "qx",
 sex       = "total")

lagrange5q0(data = input)


#2. One input and k = 0.5: q(28d,5y) (Jordan 2015)
input <- format_data(
 rate      = 0.00804138,
 lower_age = 28,
 upper_age = 365.25*5,
 type      = "qx",
 sex       = "total")

lagrange5q0(data = input, k = 0.5)


#3. Two inputs: q(0,1y) and q(1y,4y) (Belgium 1984)
input <- format_data(
 lower_age = c(0,365.25),
 upper_age = c(365.25,365.25*5),
 rate      = c(0.00996356, 0.00207788),
 type      = c("qx", "qx"),
 sex       = c("male", "male"))

lagrange5q0(data = input)

#4. Two inputs: M(0,1y) and z(28d) (Australia 1935)
 lower_age = c(0,0),
 upper_age = c(365.25, 28),
 rate      = c(0.04196866, 0.68614291),
 type      = c("mx", "zx"),
 sex       = c("total", "total"))

lagrange5q0(data = input)

#5. 22 inputs + 1 match: q(x) (Finland 1933)
data(fin1933)
fin1933$weight <- c(fin1933$n[1:22]/(365.25*5), NA)
input <- format_data(
 lower_age = fin1933$lower_age,
 upper_age = fin1933$upper_age,
 rate      = fin1933$rate,
 type      = fin1933$type,
 sex       = fin1933$sex,
 fit       = fin1933$fit,
 weight    = fin1933$weight)

lagrange5q0(data = input)

#6. One input (Model B & regional k) : q(5y) (DHS DR Congo 2013-14)
input <- format_data(
 rate      = 0.10924,
 lower_age = 0,
 upper_age = 365.25*5,
 type      = "qx",
 sex       = "total")

lagrange5q0(data = input, model = 'B', region = "Middle Africa")
}
\references{
\code{lagrange5q0} uses the  method of Langrage in order to solve the log-quadratic model presented in \url{https://doi.org/10.1215/00703370-9709538}.
}
\author{
Andrea Verhulst, \email{andrea.verhulst@ined.fr}, Julio Romero \email{Julio.Romero-Prieto@lshtm.ac.uk}
}
