% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef-data.R
\docType{data}
\name{coef}
\alias{coef}
\title{logquad coefficients}
\format{
An object of class \code{data.frame} with 132 rows and 10 columns.
}
\usage{
coef
}
\description{
logquad coefficients
}
\references{
\url{https://doi.org/10.1215/00703370-9709538}.
}
\keyword{datasets}
