% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef-data.R
\docType{data}
\name{coef}
\alias{coef}
\title{logquad coefficients}
\format{
An object of class \code{data.frame} with 66 rows and 10 columns.
}
\usage{
coef
}
\description{
logquad coefficients
}
\references{
\url{https://repository.upenn.edu/psc_publications/54/}.
}
\keyword{datasets}
