% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagrange5q0.R
\name{lagrange5q0}
\alias{lagrange5q0}
\title{lagrange5q0}
\usage{
lagrange5q0(data, k)
}
\arguments{
\item{data}{mortality inputs formated with \code{format_data}.}

\item{k}{constrained value of k (only when using a single mortality input).}
}
\value{
\code{lagrange5q0} returns a list including three elements: the predicted values of q(5y) and k, and a data frame with predicted values of q(x) and nMx.

The boundaries of the age intervals are given in days (1 year = 365.25 days, 1 month = 30.4375 days).

Different age intervals are provived for the values of q(x), i.e. the cumulative probabilities of dying between 0 and age x, and for the values nMx,  i.e. the central death rates between x and x+n.

Two types of 95\% confidence intervals (CIs) are automatically estimated: (1)	CIs based on the patterns of prediction errors observed in the underlying data when using a single mortality input (k = 0), (2) CIs based on the uncertainty of k when minimizing the RMSE over a full series of 22 q(x).
}
\description{
\code{lagrange5q0} predicts a series of 22 under-5 q(x) and nMx based on one or more mortality inputs between 0 and 5 years.
}
\details{
\code{lagrange5q0} allows four types of mortality inputs:
\enumerate{
\item A single mortality input (either nMx and nqx) for any age interval. The value of k will be assumed equal to 0 (average outcome of the model) and the value of the mortality input will be matched exactly.
\item A single mortality input (either nMx and nqx) for any age interval and a value of k. Both the mortality input and the value of k and will be matched exactly.
\item Two mortality inputs (either nMx and nqx) for any age interval. The value of k will be estimated and the value of the two mortality inputs will be matched exactly.
\item More than two mortality inputs (either nMx and nqx) for any age interval. One of them must be selected for matching. The value of k will be estimated, the selected input will be matched exactly, and the root mean square error will be minimized over the remaining mortality inputs.
}

The computation of the root mean square error (RMSE) can be weighted. When minimizing a series of q(x), i.e. cumulative probabilities of dying starting at age 0, we recommend using weights proportional to the length of the last age interval (see fourth example below).

Use \code{format_data} to prepare and verify the mortality inputs.
}
\examples{
#1. One input (k = 0): q(28d,5y) (Jordan 2015)
df <- format_data(
 rate      = 0.00804138,
 lower_age = 28,
 upper_age = 365.25*5,
 type      = "qx",
 fit       = "match",
 sex       = "total")

lagrange5q0(df)


#2. One input and k = 0.5: q(28d,5y) (Jordan 2015)
df <- format_data(
 rate      = 0.00804138,
 lower_age = 28,
 upper_age = 365.25*5,
 type      = "qx",
 fit       = "match",
 sex       = "total")

lagrange5q0(df, k = 0.5)


#3. Two inputs: q(0,1y) and q(1y,4y) (Belgium 1984)
df <- format_data(
 lower_age = c(0,365.25),
 upper_age = c(365.25,365.25*5),
 rate      = c(0.00996356, 0.00207788),
 type      = c("qx", "qx"),
 sex       = c("male", "male"))

lagrange5q0(df)


#4. 22 inputs + 1 match: q(x) (Finland 1933)
data(fin1933)
fin1933$weight <- c(fin1933$n[1:22]/(365.25*5), NA)
df <- format_data(
 lower_age = fin_1933$lower_age,
 upper_age = fin_1933$upper_age,
 rate      = fin_1933$rate,
 type      = fin_1933$type,
 sex       = fin_1933$sex,
 fit       = fin_1933$fit,
 weight    = fin_1933$weight)

lagrange5q0(df)
}
\references{
\code{lagrange5q0} uses the  method of Langrage in order to solve the log-quadratic model presented in \url{https://repository.upenn.edu/psc_publications/54/}.
}
\author{
Andrea Verhulst, \email{verhulst@sas.upenn.edu}, Julio Romero \email{Julio.Romero-Prieto@lshtm.ac.uk}
}
